src/assistant/requirements.txt
google-genai
pyaudio
openwakeword
